fix: reduce build size
Removed a dependency that was added to the workspace root by mistake.
The extra dependency was causing multiple versions of a large dependency
(Material UI) to be added many times.

SVGs for Entry were being imported directly as React components. Some of
the SVGs are very large, causing the bundle size of the compiled app to
grow to over 7 MBs. I changed the EntryLogo component to load the SVGs
from the static assets directory instead. The JavaScript bundle size was
then reduced to ~200 KB (gzip).

I re-enabled the cache for Circle CI. It was disabled as a debugging
step previously but turned out to not be the cause.

The Yarn lock file was rebuilt while clearing up the issue with the
extra dependency in the workspace root. This caused a version bump of
the type definitions for React. As a result, my custom type definitions
for the new Context feature were no longer necessary and were removed.